refactor: 패키지 구조를 수정한다
