+++ src/main/java/racingcar/domain/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.domain;
 
 @FunctionalInterface
 public interface NumberGenerator {
