+++ src/main/java/racingcar/domain/RaceNumberGenerator.java
@@ -1,7 +1,6 @@
 package racingcar.domain;
 
 import org.springframework.stereotype.Component;
-import racingcar.util.NumberGenerator;
 
 import java.util.concurrent.ThreadLocalRandom;
