+++ src/main/java/racingcar/dto/CarStatusDto.java
@@ -1,4 +1,4 @@
-package racingcar.domain.dto;
+package racingcar.dto;
 
 public class CarStatusDto {
