+++ src/main/java/racingcar/dto/RaceRequest.java
@@ -1,4 +1,4 @@
-package racingcar.domain.dto;
+package racingcar.dto;
 
 public class RaceRequest {
