+++ src/main/java/racingcar/dto/RaceResponse.java
@@ -1,4 +1,4 @@
-package racingcar.domain.dto;
+package racingcar.dto;
 
 import java.util.List;
