--- src/main/java/racingcar/exception/ConsoleException.java
@@ -1,8 +1,0 @@
-package racingcar.exception;
-
-public class ConsoleException extends GlobalException {
-
-    public ConsoleException(final String message) {
-        super(message);
-    }
-}
