--- src/main/java/racingcar/exception/DuplicateException.java
@@ -1,8 +1,0 @@
-package racingcar.exception;
-
-public class DuplicateException extends GlobalException {
-
-    public DuplicateException(final String message) {
-        super(message);
-    }
-}
