--- src/main/java/racingcar/exception/GlobalException.java
@@ -1,8 +1,0 @@
-package racingcar.exception;
-
-public class GlobalException extends RuntimeException {
-
-    public GlobalException(final String message) {
-        super(message);
-    }
-}
