--- src/main/java/racingcar/exception/NoResourceException.java
@@ -1,8 +1,0 @@
-package racingcar.exception;
-
-public class NoResourceException extends GlobalException {
-
-    public NoResourceException(String message) {
-        super(message);
-    }
-}
