+++ src/main/resources/static/css/styles.css
@@ -2,11 +2,13 @@ body {
     font-family: Arial, sans-serif;
     background-color: #f2f2f2;
 }
+
 h1 {
     margin: 20px 0;
     font-size: 24px;
     text-align: center;
 }
+
 /* 입력 영역 */
 .input-area {
     display: flex;
@@ -85,11 +87,13 @@ table {
     max-width: 800px;
     width: 100%;
 }
+
 th, td {
     border: 1px solid #ccc;
     padding: 8px;
     text-align: center;
 }
+
 th {
     background-color: #f2f2f2;
     font-weight: bold;
